# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DARK TIERS Minecraft PvP Ranking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tiers:
    get:
      operationId: listTiers
      tags: [tiers]
      summary: List all tiers
      parameters:
        - name: gamemodeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tiers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tier"
    post:
      operationId: createTier
      tags: [tiers]
      summary: Create a tier (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TierInput"
      responses:
        "201":
          description: Created tier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tier"
  /tiers/{id}:
    patch:
      operationId: updateTier
      tags: [tiers]
      summary: Update a tier (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TierUpdate"
      responses:
        "200":
          description: Updated tier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tier"
    delete:
      operationId: deleteTier
      tags: [tiers]
      summary: Delete a tier (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PLAYERS
...
components:
  schemas:
...
    Tier:
      type: object
      required: [id, name, slug, rank, minRating, maxRating, color]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        rank:
          type: integer
        minRating:
          type: integer
        maxRating:
          type: ["integer", "null"]
        color:
          type: string
        gamemodeId:
          type: ["integer", "null"]
        createdAt:
          type: string
    TierInput:
      type: object
      required: [name, slug, rank, minRating, color]
      properties:
        name:
          type: string
        slug:
          type: string
        rank:
          type: integer
        minRating:
          type: integer
        maxRating:
          type: integer
        color:
          type: string
        gamemodeId:
          type: integer
    TierUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        rank:
          type: integer
        minRating:
          type: integer
        maxRating:
          type: integer
        color:
          type: string
...