# lib/api-zod/src/generated/types/tierInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DARK TIERS Minecraft PvP Ranking Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TierInput {
  name: string;
  slug: string;
  rank: number;
  minRating: number;
  maxRating?: number;
  color: string;
  gamemodeId?: number;
}